controller/src/requirements.txt
mysql-connector-python
opencv-python-headless
rktellolib==1.0.2
requests